added multiple test cases in cypress
