--- cypress.json
@@ -1,3 +1,0 @@
-{
-  "baseUrl": "http://localhost:3000"
-}
